Move device setup into utils.device
